refactor(eve): expose result-binding queries from handles/query
findRunningAgentHandle, isResultBoundToRunningHandle, and
isInboxResultFromRunningHandle read only the handle store;
runtime-actions.ts was just their first caller. Move them next to the
store they query so turn-workflow and future consumers import the
binding rules from the handles API instead of the harness internals.
