+++ packages/eve/src/harness/runtime-actions.test.ts
@@ -1,9 +1,11 @@
+import {
+  isInboxResultFromRunningHandle,
+  isResultBoundToRunningHandle,
+} from "#harness/handles/query.js";
 import { describe, expect, it } from "vitest";
 
 import {
   getPendingRuntimeActionBatch,
-  isInboxResultFromRunningHandle,
-  isResultBoundToRunningHandle,
   resolvePendingRuntimeActions,
   resolveToolCallInputObject,
   setPendingRuntimeActionBatch,
